refactor: Cloth 클래스 분리
Cloth 클래스 안에 있던 기능들을 DyeManager 클래스로 분리하였다
