+++ Assets/PrivateFolder/HyunJune/Prefab/ColorMission.prefab
@@ -87,6 +87,7 @@ GameObject:
   - component: {fileID: 4667797120796771301}
   - component: {fileID: 4667797120796771307}
   - component: {fileID: 4667797120796771306}
+  - component: {fileID: 3913278387886380554}
   m_Layer: 5
   m_Name: ColorMission
   m_TagString: Untagged
@@ -156,6 +157,19 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &3913278387886380554
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4667797120796771300}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: def4413a07a2bb947bc86619351614ba, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  cloth: {fileID: 1939753025}
 --- !u!1 &4667797120980209980
 GameObject:
   m_ObjectHideFlags: 0
@@ -323,6 +337,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ce254a8a87ba31040b4a2132d8df5a17, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  clothImage: {fileID: 0}
   curColor: 0
 --- !u!61 &3390831340103947761
 BoxCollider2D:
