+++ Assets/PrivateFolder/HyunJune/Script/DyeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: def4413a07a2bb947bc86619351614ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
